refactor(character-builder): move lineage picker from Basics to Origin
Lineage choice (dragonborn ancestry, elf subrace, etc.) is a species
sub-choice and belongs in the Origin step alongside background sub-choices,
not in Basics. Extracted to its own LineagePicker.tsx so both ergonomics
and testing are simpler.

Tests for lineage behavior moved from BasicsStep.test.tsx to OriginStep.test.tsx.
BasicsStep retains a single negative test confirming the picker no longer
renders there.
